Creacion del checkout

+++ zuzuPlayDjango/rest_cart/views.py
@@ -1,9 +1,9 @@
 from rest_framework import status
-from rest_framework.decorators import api_view, permission_classes
+from rest_framework.decorators import api_view
 from rest_framework.response import Response
-from rest_framework.permissions import IsAuthenticated
 from django.views.decorators.csrf import csrf_exempt
 from .models import Carrito, CarritoProducto
+from orders.models import Direccion
 from .serializers import CarritoSerializer, CarritoProductoSerializer
 
 
@@ -13,13 +13,20 @@ from .serializers import CarritoSerializer, CarritoProductoSerializer
 def carrito(request):
     if request.method == 'GET':
         if request.user.is_authenticated:
-            carrito = Carrito.objects.get_or_create(usuario_id=request.user.id)
+            carrito = Carrito.objects.filter(usuario=request.user)
+            if carrito.count() == 0:
+                try:
+                    direccion = Direccion.objects.get(usuario=request.user, principal=True)
+                except Direccion.DoesNotExist:
+                    direccion = None
+                carrito = Carrito.objects.create(usuario=request.user, direccion=direccion)
         else:
             if not request.session or not request.session.session_key:
                 request.session.save()
             carrito = Carrito.objects.get_or_create(sesion_id=request.session.session_key)
         serializer = CarritoSerializer(carrito, many=True)
         return Response(serializer.data)
+
 
 @csrf_exempt
 @api_view(['POST'])
@@ -35,14 +42,22 @@ def agregar_al_carrito(request):
         else:
             serializer = CarritoProductoSerializer(data=request.data)
         if serializer.is_valid():
+            oldTotal = 0
             if producto is not None:
                 serializer.validated_data['cantidad'] = producto.cantidad + int(request.data['cantidad'])
+                oldTotal = producto.total
             serializer.validated_data['total'] = \
                 serializer.validated_data['producto'].precioProducto * serializer.validated_data['cantidad']
-            serializer.save()
-            return Response(serializer.data, status=status.HTTP_201_CREATED)
+            subtotal = update_cart_subtotal(request, serializer.validated_data['carrito'].id,
+                                            serializer.validated_data['total'], oldTotal)
+            if subtotal:
+                serializer.save()
+                return Response(serializer.data, status=status.HTTP_201_CREATED)
+            else:
+                return Response(status=status.HTTP_400_BAD_REQUEST)
         else:
             return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+
 
 @csrf_exempt
 @api_view(['GET', 'PUT', 'DELETE'])
@@ -60,10 +75,35 @@ def modificar_carrito(request, id_carrito, id_producto):
         if serializer.is_valid():
             serializer.validated_data['total'] = \
                 serializer.validated_data['producto'].precioProducto * serializer.validated_data['cantidad']
-            serializer.save()
-            return Response(serializer.data)
+            if update_cart_subtotal(request, producto.carrito_id,
+                                    serializer.validated_data['total'], producto.total):
+                serializer.save()
+                return Response(serializer.data)
+            else:
+                return Response(status=status.HTTP_400_BAD_REQUEST)
         else:
             return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
     elif request.method == 'DELETE':
-        producto.delete()
-        return Response(status=status.HTTP_204_NO_CONTENT)
+        if update_cart_subtotal(request, producto.carrito_id, producto.total):
+            producto.delete()
+            return Response(status=status.HTTP_204_NO_CONTENT)
+        else:
+            return Response(status=status.HTTP_400_BAD_REQUEST)
+
+
+def update_cart_subtotal(request, cartId, total, oldTotal=0):
+    try:
+        carrito = Carrito.objects.get(id=cartId)
+    except Carrito.DoesNotExist:
+        return False
+    if request.method == 'DELETE':
+        subtotal = carrito.subtotal - total
+    else:
+        subtotal = carrito.subtotal - oldTotal + total
+    data = {'id': cartId, 'subtotal': subtotal, 'total': subtotal}
+    serializer = CarritoSerializer(carrito, data=data)
+    if serializer.is_valid():
+        serializer.save()
+        return True
+    else:
+        return False
